+++ documentation/docs/changelog.md
@@ -6,7 +6,8 @@ sidebar_position: 2
 
 ## yyyy.mm.dd - version 1.12.0
 
-* Add the [Interval:startImmediately](resources/support/interval#usage) method
+* Add the [Interval:startImmediately()](resources/support/interval#usage) method
+* Add the [Window:toggleVisibility()](resources/views/window#showing-and-hiding-the-window) method
 * Fix a couple of doc tag namespaces in the Facades package
 
 ## 2024.08.01 - version 1.11.0